fix error

+++ toolbox/dataset_monai.py
@@ -69,7 +69,7 @@ class RSNA_MICCAI_MONAI_Dataset(pl.LightningDataModule):
             raise ValueError(f"Invalid mode {mode}")
 
         csv_file = "train_labels.csv" if mode == "train" else "sample_submission.csv"
-        df = pd.read_csv(os.path.join(self.config.dataset_dir, csv_file), dtype={"BraTS21ID": str, "MGMT_value": int})
+        df = pd.read_csv(os.path.join(self.config.dataset_dir, csv_file), dtype={"BraTS21ID": str, "MGMT_value": float})
         subject_dict_list = []
         for _, row in df.iterrows():
             subject_id = row["BraTS21ID"]
